Pull webpack actions out of Neutrino class.
